Correct type for curl_easy_getinfo, which could fail on some compilers.

+++ arbiter/util/http.cpp
@@ -215,7 +215,7 @@ Response Curl::get(
         Query query,
         const std::size_t reserve)
 {
-    int httpCode(0);
+    long httpCode(0);
     std::vector<char> data;
 
     if (reserve) data.reserve(reserve);
@@ -245,7 +245,7 @@ Response Curl::get(
 
 Response Curl::head(std::string path, Headers headers, Query query)
 {
-    int httpCode(0);
+    long httpCode(0);
     std::vector<char> data;
 
     init(path, headers, query);
@@ -283,7 +283,7 @@ Response Curl::put(
     init(path, headers, query);
     if (m_verbose) curl_easy_setopt(m_curl, CURLOPT_VERBOSE, 1L);
 
-    int httpCode(0);
+    long httpCode(0);
 
     std::unique_ptr<PutData> putData(new PutData(data));
 
@@ -325,7 +325,7 @@ Response Curl::post(
     init(path, headers, query);
     if (m_verbose) curl_easy_setopt(m_curl, CURLOPT_VERBOSE, 1L);
 
-    int httpCode(0);
+    long httpCode(0);
 
     std::unique_ptr<PutData> putData(new PutData(data));
     std::vector<char> writeData;
